chore(select-field): move to packages
