+++ packages/components/fields/select-field/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './select-field';
+
+export { default as version } from './version';
