--- src/components/fields/select-field/index.js
@@ -1,1 +1,0 @@
-export { default } from './select-field';
